Revert "Feature/ReferencePushPullFeeder (#977)"
This reverts commit ce654ca8c47c022605ca1706cd9b629c068640fd.
