+++ src/main/java/org/openpnp/gui/components/CameraView.java
@@ -1280,7 +1280,10 @@ public class CameraView extends JComponent implements CameraListener {
         }
     }
 
-    public Location getCameraViewCenterOffsetsFromXy(int x, int y) {
+    private void moveToClick(MouseEvent e) {
+        int x = e.getX();
+        int y = e.getY();
+
         // Find the difference in X and Y from the center of the image
         // to the mouse click.
         double offsetX = (scaledWidth / 2.0D) - (x - imageX);
@@ -1294,17 +1297,12 @@ public class CameraView extends JComponent implements CameraListener {
         offsetX *= scaledUnitsPerPixelX;
         offsetY *= scaledUnitsPerPixelY;
 
-        // The offsets now represent the distance in the Camera's units per pixel's units.
+        // The offsets now represent the distance to move the camera
+        // in the Camera's units per pixel's units.
 
         // Create a location in the Camera's units per pixel's units
         // and with the values of the offsets.
         Location offsets = camera.getUnitsPerPixel().derive(offsetX, offsetY, 0.0, 0.0);
-        return offsets;
-    }
-
-    private void moveToClick(MouseEvent e) {
-        // Get the offset from the Camera view center in Camera's units.
-        Location offsets = getCameraViewCenterOffsetsFromXy(e.getX(), e.getY());
         // And move there.
         UiUtils.submitUiMachineTask(() -> {
             if (camera.getHead() == null) {
