+++ src/main/java/org/openpnp/machine/reference/feeder/BlindsFeeder.java
@@ -31,7 +31,9 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import javax.swing.Action;
 
@@ -70,7 +72,6 @@ import org.openpnp.util.TravellingSalesman;
 import org.openpnp.util.VisionUtils;
 import org.openpnp.vision.FluentCv;
 import org.openpnp.vision.Ransac.Line;
-import org.openpnp.vision.SimpleHistogram;
 import org.openpnp.vision.pipeline.CvPipeline;
 import org.pmw.tinylog.Logger;
 import org.simpleframework.xml.Attribute;
@@ -514,6 +515,77 @@ public class BlindsFeeder extends ReferenceFeeder {
             return angle;
         }
 
+        /** 
+         * Very simple histogram with zero knowledge about the range. 
+         * Good only for low numbers of measurements.
+         * The bins are filled with an 2-bin wide square kernel centered on the measurement, recording 
+         * the overlap of the kernel with the bins. If the distribution of measurements clusters near the 
+         * edge between two bins (dithering between the two), the recorded weight is still comparable to a 
+         * similar distribution centered in the middle of a bin. Therefore the minimum/maximum is 
+         * still reliably captured in these cases, even in the presence of background noise. 
+         */
+        class Histogram {
+            private Map<Integer, Double> histogram = new HashMap<>();
+            private Integer minimum = null;
+            private double minimumVal = Double.NaN;
+            private Integer maximum = null;
+            private double maximumVal = Double.NaN;
+            private double resolution;
+            public Histogram(double resolution) {
+                this.resolution = resolution;
+            }
+            void add(double key, double val) {
+                // the three histogram bins around the key position
+                int binY1 = (int)Math.round(key/resolution);   
+                int binY0 = binY1-1; 
+                int binY2 = binY1+1;
+                // calculate the weight according to overlap with the kernel spreading 2 bins 
+                double w1 = 1.0;
+                double w0 = ((binY0+1.5)*resolution - key)/resolution;
+                double w2 = (-(binY2-1.5)*resolution + key)/resolution;
+                add(binY1, w1*val);
+                add(binY0, w0*val);
+                add(binY2, w2*val);
+            }
+            void add(int key, double val) {
+
+                double newVal = histogram.get(key) == null ? val : histogram.get(key)+val;
+                histogram.put(key, newVal);
+                if (maximum == null || maximumVal < newVal) {
+                    maximumVal = newVal;
+                    maximum = key;
+                }
+                if (minimum == null || minimumVal > newVal) {
+                    minimumVal = newVal;
+                    minimum = key;
+                }
+            }
+            public Map<Integer, Double> getHistogram() {
+                return histogram;
+            }
+            public Integer getMinimum() {
+                return minimum;
+            }
+            public double getMinimumKey() {
+                return minimum == null ? Double.NaN : minimum*resolution;
+            }
+            public double getMinimumVal() {
+                return minimumVal;
+            }
+            public Integer getMaximum() {
+                return maximum;
+            }
+            public double getMaximumKey() {
+                return maximum == null ? Double.NaN : maximum*resolution;
+            }
+            public double getMaximumVal() {
+                return maximumVal;
+            }
+            public double getResolution() {
+                return resolution;
+            }
+        }
+
         @SuppressWarnings("unchecked")
         public FindFeatures invoke() throws Exception {
             List<RotatedRect> results = null;
@@ -566,8 +638,8 @@ public class BlindsFeeder extends ReferenceFeeder {
                 fiducials = new ArrayList<>();
 
                 // Create center and corner histograms
-                SimpleHistogram histogramUpper = new SimpleHistogram(0.1);
-                SimpleHistogram histogramLower = new SimpleHistogram(0.1);
+                Histogram histogramUpper = new Histogram(0.1);
+                Histogram histogramLower = new Histogram(0.1);
 
                 for (RotatedRect result : results) {
                     org.opencv.core.Point points[] = new org.opencv.core.Point[4];
@@ -704,7 +776,7 @@ public class BlindsFeeder extends ReferenceFeeder {
 
                 // Try to determine the pocket pitch by creating a histogram of sorted blinds distances. 
                 Location previous = null;
-                SimpleHistogram histogramPitch = new SimpleHistogram(2);
+                Histogram histogramPitch = new Histogram(2);
                 for (RotatedRect rect : blinds) {
                     Location location = transformMachineToFeederLocation(VisionUtils.getPixelLocation(camera, rect.center.x, rect.center.y))
                             .convertToUnits(LengthUnit.Millimeters);
@@ -722,7 +794,7 @@ public class BlindsFeeder extends ReferenceFeeder {
                 double pocketPitchPosMm = (getPocketPitch().getValue() != 0. ? 
                         getPocketPitch().convertToUnits(LengthUnit.Millimeters).getValue() 
                         :   pocketPitchMm); 
-                SimpleHistogram histogramDistance = new SimpleHistogram(0.05);
+                Histogram histogramDistance = new Histogram(0.05);
                 for (RotatedRect rect : blinds) {
                     Location location = transformMachineToFeederLocation(VisionUtils.getPixelLocation(camera, rect.center.x, rect.center.y))
                             .convertToUnits(LengthUnit.Millimeters);
@@ -768,10 +840,8 @@ public class BlindsFeeder extends ReferenceFeeder {
                     drawRotatedRects(resultMat, getFiducials(), Color.white);
                     drawLines(resultMat, getLines(), new Color(0, 0, 128));
                     drawPartNumbers(resultMat, Color.orange);
-                    if (Logger.getLevel() == org.pmw.tinylog.Level.DEBUG || Logger.getLevel() == org.pmw.tinylog.Level.TRACE) {
-                        File file = Configuration.get().createResourceFile(getClass(), "blinds-feeder", ".png");
-                        Imgcodecs.imwrite(file.getAbsolutePath(), resultMat);
-                    }
+                    File file = Configuration.get().createResourceFile(getClass(), "blinds-feeder", ".png");
+                    Imgcodecs.imwrite(file.getAbsolutePath(), resultMat);
                     BufferedImage showResult = OpenCvUtils.toBufferedImage(resultMat);
                     resultMat.release();
                     MainFrame.get().getCameraViews().getCameraView(camera)
@@ -1134,16 +1204,14 @@ public class BlindsFeeder extends ReferenceFeeder {
     public static class NozzleAndTipForPushing {
         private Nozzle nozzle;
         private NozzleTip nozzleTipLoadedBefore;
-        private boolean changed;
 
-        public NozzleAndTipForPushing(Nozzle nozzle, NozzleTip nozzleTipLoadedBefore, boolean changed) {
+        public NozzleAndTipForPushing(Nozzle nozzle, NozzleTip nozzleTipLoadedBefore) {
             super();
             this.nozzle = nozzle;
             this.nozzleTipLoadedBefore = nozzleTipLoadedBefore;
-            this.changed = changed;
         }
         void restoreNozzleTipLoadedBefore() throws Exception {
-            if (this.nozzle != null & this.nozzleTipLoadedBefore != null && this.changed) {
+            if (this.nozzle != null & this.nozzleTipLoadedBefore != null) {
                 if (this.nozzleTipLoadedBefore != this.nozzle.getNozzleTip()) {
                     this.nozzle.loadNozzleTip(this.nozzleTipLoadedBefore);
                 }
@@ -1161,9 +1229,6 @@ public class BlindsFeeder extends ReferenceFeeder {
         public NozzleTip getNozzleTipLoadedBefore() {
             return nozzleTipLoadedBefore;
         }
-        public boolean isChanged() {
-            return changed;
-        }
     }
     public static NozzleAndTipForPushing getNozzleAndTipForPushing(Part favoredCompatiblePart, boolean loadNozzleTipIfNeeded) throws Exception {
         // first search for any NozzleTip already loaded that may be used for pushing 
@@ -1180,9 +1245,9 @@ public class BlindsFeeder extends ReferenceFeeder {
                             if (favoredCompatiblePart.getPackage()
                                     .getCompatibleNozzleTips().contains(nozzleTip)) {
                                 // Return the nozzle and nozzle tip.
-                                return new NozzleAndTipForPushing(nozzle, nozzleTip, false);
+                                return new NozzleAndTipForPushing(nozzle, nozzleTip);
                             }
-                        } 
+                        }
                     }
                 }
             }
@@ -1193,7 +1258,7 @@ public class BlindsFeeder extends ReferenceFeeder {
                     NozzleTip nozzleTip = nozzle.getNozzleTip();
                     if (nozzleTip.isPushAndDragAllowed()) {
                         // Return the nozzle and nozzle tip.
-                        return new NozzleAndTipForPushing(nozzle, nozzleTip, false);
+                        return new NozzleAndTipForPushing(nozzle, nozzleTip);
                     }
                 }
             }
@@ -1211,7 +1276,7 @@ public class BlindsFeeder extends ReferenceFeeder {
                                 // Alas, found one for pushing, load it
                                 nozzle.loadNozzleTip(pushNozzleTip);
                                 // And return the nozzle and nozzle tip.
-                                return new NozzleAndTipForPushing(nozzle, nozzleTip, true);
+                                return new NozzleAndTipForPushing(nozzle, nozzleTip);
                             }
                         }
                     }
@@ -1221,47 +1286,27 @@ public class BlindsFeeder extends ReferenceFeeder {
             throw new Exception("BlindsFeeder: No empty Nozzle/NozzleTip found that allows pushing.");
         }
         // None compatible.
-        return new NozzleAndTipForPushing(null, null, false);
+        return new NozzleAndTipForPushing(null, null);
     }
 
-
     @Override
-    public Location getJobPreparationLocation() {
+    public void prepareForJob(List<Feeder> feedersToPrepare) throws Exception {
+        super.prepareForJob(feedersToPrepare);
         if (getCoverActuation() == CoverActuation.OpenOnJobStart
                 && ! isCoverOpen()) {
-            return getUncalibratedPickLocation(0);
-        }
-        else {
-            return null;
-        }
-    }
-
-    // transient store for the job preparation nozzle tip change
-    private NozzleAndTipForPushing nozzleAndTipForPushing = null;
-
-    @Override
-    public void prepareForJob(boolean visit) throws Exception {
-        super.prepareForJob(visit);
-        if (visit && getCoverActuation() == CoverActuation.OpenOnJobStart
-                && ! isCoverOpen()) {
-            // Get the push nozzle for the current position. This will also throw if none is allowed.
-            // Note, we save this for pass two of the feeder prep to restore.
-            nozzleAndTipForPushing = BlindsFeeder.getNozzleAndTipForPushing(null, true);
-
-            // Actuate the cover.
-            actuateCover(true);
-            
-            if (!nozzleAndTipForPushing.isChanged()) {
-                // no need to rememeber
-                nozzleAndTipForPushing = null;
-            }
-        }
-        else if (! visit) {
-            // in the non-visit preparation step, let's restore the nozzle tip if changed before
-            if (nozzleAndTipForPushing != null) {
-                nozzleAndTipForPushing.restoreNozzleTipLoadedBefore();
-                nozzleAndTipForPushing = null;
-            }
+            // Note, we will not only open our own cover, but all the BlindFeeders' covers in bulk
+            // to allow for the TravellingSalesman optimization.
+            // prepareForJob() will be called for subsequent BlindsFeeders, but because the covers 
+            // are already registered as open, there will be no further action. 
+            List<BlindsFeeder> feederList = getFeedersWithCoverToActuate(feedersToPrepare, 
+                    new CoverActuation[] { CoverActuation.OpenOnJobStart }, 
+                    true); 
+            // Now bulk-open the covers. 
+            // If needed, load a NozzleTip that allows pushing and then restore the previous NozzleTip.
+            // The restore may result in one unnecessary change but the alternative is unexpected behaviour
+            // as the planner will take the currently loaded nozzle tips into consideration and therefore 
+            // favor the pushing nozzle tip, which might be completely unwanted.
+            actuateListedFeederCovers(feederList, true, false);
         }
     }
 
@@ -1740,7 +1785,7 @@ public class BlindsFeeder extends ReferenceFeeder {
             pipeline.setProperty("camera", camera);
             pipeline.setProperty("feeder", this);
 
-            /* TODO: read the override property in the FilterContours stage
+            /* this won't work, properties must be hardwired in the stages
              * 
              * // Provide pixel min/max area to pipeline.
             // We restrict this to 24mm tape carrier having a 20.1mm max pocket size.
@@ -2020,6 +2065,7 @@ public class BlindsFeeder extends ReferenceFeeder {
     public void setFeedCount(int feedCount) {
         int oldValue = this.feedCount;
         this.feedCount = feedCount;
+        //this.visionOffsets = null;
         firePropertyChange("feedCount", oldValue, feedCount);
     }
