+++ src/main/java/org/openpnp/model/Package.java
@@ -42,9 +42,6 @@ public class Package extends AbstractModelObject implements Identifiable {
     private String description;
 
     @Attribute(required = false)
-    private String tapeSpecification;
-
-    @Attribute(required = false)
     private double pickVacuumLevel;
 
     @Attribute(required = false)
@@ -92,16 +89,6 @@ public class Package extends AbstractModelObject implements Identifiable {
         Object oldValue = this.description;
         this.description = description;
         firePropertyChange("description", oldValue, description);
-    }
-
-    public String getTapeSpecification() {
-        return tapeSpecification;
-    }
-
-    public void setTapeSpecification(String tapeSpecification) {
-        Object oldValue = this.tapeSpecification;
-        this.tapeSpecification = tapeSpecification;
-        firePropertyChange("tapeSpecification", oldValue, tapeSpecification);
     }
 
     public void setPlaceBlowOffLevel(double level) {
