+++ src/main/java/org/openpnp/spi/base/AbstractFeeder.java
@@ -67,9 +67,7 @@ public abstract class AbstractFeeder extends AbstractModelObject implements Feed
 
     @Override
     public void setPart(Part part) {
-        Object oldValue = this.part;
         this.part = part;
-        firePropertyChange("part", oldValue, part);
         this.partId = part.getId();
     }
